requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.9.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-ninja==1.4.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.13
fonttools==4.57.0
google-auth==2.40.3
gunicorn==23.0.0
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
nh3==0.2.21
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pyloco==0.0.139
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-fuzzy==0.5.0
scipy==1.15.2
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
tornado==6.4.2
twine==6.1.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.3.0
ushlex==0.99.1
websocket-client==1.8.0
